feat: Removed Angular Material and added Bootstrap
This commit removes Angular Material from the project and replaces it with Bootstrap 5.3.0-alpha1. Angular Material is a UI component library for Angular applications, while Bootstrap is a popular CSS framework that offers a wide range of styling and component options.

By removing Angular Material and adding Bootstrap, we are making a significant change to the project's visual design and functionality. Bootstrap offers a range of benefits such as responsive design, pre-built CSS classes, and a large community of developers for support.

The version added, Bootstrap 5.3.0-alpha1, is an alpha version of Bootstrap 5.3.0, which means it's still in the testing phase and may contain bugs or incomplete features. However, we believe that it will offer better performance and functionality than the previous Angular Material library, and we're excited to see the improvements it brings to the project.